requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
certifi==2017.7.27.1      # via python-telegram-bot
future==0.16.0            # via python-telegram-bot
olefile==0.44             # via pillow
peewee==2.10.1
pillow==4.2.1
python-telegram-bot==8.0
